requirements.txt
flask>=2.0.0
pypresence>=4.0.0
ytmusicapi>=1.0.0
python-dotenv>=1.0.0
